refactor: extract TypeVarEmitter.bridgeFromKind (Phase 5, design §3.6)
Add a shared TypeVarEmitter object in scalus.compiler.sir.lowering.typegens
holding bridgeFromKind — the Transparent/Unwrapped/Fixed dispatch for
TypeVar-source values, parameterized by the input's typegen via lctx
lookups (so it works for any source typegen, not just DataConstrEmitter).

SumUplcConstrEmitter's previous local bridgeFromTypeVar (added in
aa2079a56) is replaced by a call to TypeVarEmitter.bridgeFromKind.

Note: SumDispatch.sumCaseImpl and SumListEmitterCommon.emitConvert keep
their open-coded TypeVar arms for now — their hardcoded reprs (DataConstr
/ PackedSumDataList) preserve subtle differences from what
lctx.typeGenerator(...).default* returns for some typegens. Migrating
those sites needs a per-arm semantic audit; deferred to a future commit.
